--- src/main/java/guru/springframework/springrecipe5app/exceptions/NumberFormatException.java
@@ -1,20 +1,0 @@
-package guru.springframework.springrecipe5app.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class NumberFormatException extends RuntimeException {
-
-    public NumberFormatException() {
-        super();
-    }
-
-    public NumberFormatException(String message) {
-        super(message);
-    }
-
-    public NumberFormatException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
